+++ conjur-env/build.sh
@@ -6,3 +6,4 @@ rm -rf ../vendor/conjur-env
 
 docker compose build
 docker compose run --rm conjur-env-builder
+docker compose run --rm conjur-win-env-builder